pyproject.toml
[project]
name = "psychobench-pro"
version = "0.1.0"
description = "Benchmarking LLMs' Psychological Portray (PsychoBench)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.0",
    "hydra-joblib-launcher",
    "numpy>=1.21.0",
    "scipy>=1.9.0",
    "matplotlib>=3.5.0",
    "pandas>=1.3.0",
    "requests>=2.28.0",
    "litellm[google]>=1.0.0",
    "tenacity>=8.0.0",
    "tqdm>=4.64.0",
    "loguru>=0.7.0",
    "anthropic>=0.39.0",
    "pydantic>=2.0.0",
    "tiktoken>=0.5.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"utils.py" = ["E501"]  # Ignore line length for data structures like payload_template

[tool.ruff.lint.isort]
known-first-party = ["psychobench"]
